chore/#29 : profile,select-template 뷰 AsyncBoundary로 감싸고 폴더구조 변경
